@@ -14,16 +14,16 @@
   "dependencies": {
     "@angular-eslint/eslint-plugin": "^18.4.0",
     "@angular-eslint/schematics": "^18.4.0",
-    "@angular/animations": "^17.3.12",
-    "@angular/cdk": "^16.2.14",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^16.2.12",
-    "@angular/forms": "^17.3.12",
+    "@angular/animations": "^18.2.8",
+    "@angular/cdk": "^17.3.10",
+    "@angular/common": "^18.2.8",
+    "@angular/compiler": "^18.2.8",
+    "@angular/core": "^18.2.8",
+    "@angular/forms": "^18.2.8",
     "@angular/material": "^16.2.14",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
+    "@angular/platform-browser": "^18.2.8",
+    "@angular/platform-browser-dynamic": "^18.2.8",
+    "@angular/router": "^18.2.8",
     "@geoman-io/leaflet-geoman-free": "^2.17.0",
     "@types/geojson": "^7946.0.14",
     "@types/leaflet": "^1.9.12",
@@ -34,12 +34,12 @@
     "ngx-toastr": "^19.0.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.13.3"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^17.3.10",
     "@angular/cli": "~17.3.10",
-    "@angular/compiler-cli": "^17.3.12",
+    "@angular/compiler-cli": "^18.2.8",
     "@types/google.maps": "^3.58.1",
     "@types/jasmine": "~4.3.0",
     "@types/jquery": "^3.5.30",
